Enhance GCS Fully-Qualified Name (FQN) handling
